Fixed ban function circluar dep
