+++ src/views/Moderator/Moderator.tsx
@@ -16,19 +16,16 @@
  */
 
 import * as React from "react";
+import * as moment from "moment";
+import * as data from "data";
 import { Link } from "react-router-dom";
 import { _ } from "translate";
-import { post, put } from "requests";
 import { PaginatedTable } from "PaginatedTable";
 import { Card } from "material";
 import { SearchInput } from "misc-ui";
-import { BanModal } from "BanModal";
-import { openModal } from "Modal";
 import { Player } from "Player";
-import * as moment from "moment";
 import { chat_markup } from "Chat";
-import * as data from "data";
-import swal from "sweetalert2";
+import { ban, shadowban } from "./ban_functions";
 
 const whitelist = [
     "aol.com",
@@ -606,68 +603,6 @@ export class ColorTableToggle extends React.Component<{}, any> {
                 </label>
             </div>
         );
-    }
-}
-
-function moderate(player_id, prompt, obj) {
-    return new Promise((resolve, reject) => {
-        swal({
-            text: prompt,
-            input: "text",
-            showCancelButton: true,
-        }).then((reason) => {
-            obj.moderation_note = reason;
-            console.log(obj);
-            put("players/" + player_id + "/moderate", obj)
-                .then(resolve)
-                .catch(reject);
-        }, reject);
-    });
-}
-export function ban(player_id) {
-    if (player_id < 0) {
-        return post("moderation/shadowban_anonymous_user", {
-            ban: 1,
-            user_id: player_id,
-        });
-    } else {
-        /*
-        return moderate(player_id, "Reason for banning? This will be visible to the player now.", {
-            is_banned: 1,
-        });
-        */
-        openModal(<BanModal player_id={player_id} />);
-    }
-    return undefined;
-}
-export function shadowban(player_id) {
-    if (player_id < 0) {
-        return post("moderation/shadowban_anonymous_user", {
-            ban: 1,
-            user_id: player_id,
-        });
-    } else {
-        return moderate(player_id, "Reason for shadow banning?", { is_shadowbanned: 1 });
-    }
-}
-export function remove_ban(player_id) {
-    if (player_id < 0) {
-        return post("moderation/shadowban_anonymous_user", {
-            ban: 0,
-            user_id: player_id,
-        });
-    } else {
-        return moderate(player_id, "Reason for removing ban?", { is_banned: 0 });
-    }
-}
-export function remove_shadowban(player_id) {
-    if (player_id < 0) {
-        return post("moderation/shadowban_anonymous_user", {
-            ban: 0,
-            user_id: player_id,
-        });
-    } else {
-        return moderate(player_id, "Reason for removing the shadow ban?", { is_shadowbanned: 0 });
     }
 }
